+++ idasen_controller/desk.py
@@ -25,21 +25,22 @@ class Desk:
         print("Capabilities: {}".format(capabilities))
 
         # Read the user id
-        user_id = (await DPGService.dpg_command(client, DPGService.DPG.CMD_USER_ID))
+        user_id = await DPGService.dpg_command(client, DPGService.DPG.CMD_USER_ID)
         print("User ID: {}".format(bytes_to_hex(user_id)))
         if user_id and user_id[0] != 1:
-            # Set user_id to more privileged
+            # For DPG1C it is important that the first byte is set to 1
+            # The other bytes do not seem to matter 
             user_id[0] = 1
             print("Setting user ID to {}".format(bytes_to_hex(user_id)))
             await DPGService.dpg_command(client, DPGService.DPG.CMD_USER_ID, user_id)
 
         # Check if base height should be taken from controller
-        if config.base_height == 0:
-            resp = (await DPGService.dpg_command(client, DPGService.DPG.CMD_BASE_OFFSET))
+        if config.base_height == None:
+            resp = await DPGService.dpg_command(client, DPGService.DPG.CMD_BASE_OFFSET)
             if resp:
-                base_height = struct.unpack("<H", resp[1:])[0]
-                print("base_height taken from controller {}".format(base_height))
-                config.base_height = base_height/10
+                base_height = struct.unpack("<H", resp[1:])[0] / 10
+                print("Base height from desk: {:4.0f}mm".format(base_height))
+                config.base_height = base_height